Moved out error message into a generic Messages handler
using a messageBus using EventEmitter from events.

As ErrorAlert was no longer necessary it was removed but this lead to an
interesting regression in running tests going from 1:20 to almost 8
minutes. This was due to a recompilation occuring every time a capyabara
visit occuring and the fact that the experimental/ExProfile/index.jsx
still required and rendered it. As no test hits experimental no tests
broke but node and weback took a long time to search for the now missing
component without any output. Presumably the output in
public/packs-test/js/experimental_app*.js would have some clues.
Also this does render in the test server logs

  ERROR in ./app/javascript/components/experimental/ExProfile/index.jsx
  Module not found: Error: Can't resolve '../../ErrorAlert' ...
  @ ./app/javascript/components/experimental/ExProfile/index.jsx
  @ ./app/javascript/packs/experimental_app.js

Co-authored-by: Dave McDonald <davesaddress.nz@gmail.com>